backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
transformers==4.35.0
torch>=2.0.0
vaderSentiment==3.3.2
scikit-learn==1.3.2
numpy>=1.24.0
pandas>=2.0.0
httpx==0.25.2
